requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.21
tenacity==8.2.3
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.7
xyzservices==2023.7.0
